// frontend/src/core/models/pilatDictionaries.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:43:35 GMT-0400 (Bolivia Time)
 * Time: 2:43:35
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:43:35 GMT-0400 (Bolivia Time)
 * Last time updated: 2:43:35
 *
 * Caution: es-sections will be replaced by script execution
 */

//<es-section>

import {PilatParams} from "./pilatParams";

//</es-section>

export class PilatDictionaries {

  //<es-section>
  
  _id: string;
  
  
  id: number;
  
  
  
  
  dic_code: string;
  
  dic_description: string;
  
  dic_group: string;
  
  createdBy: string;
  
  updatedBy: string;
  
  
  
  dic_par_status_id: string;
  
  
  dueAt: Date;
  
  createdAt: Date;
  
  updatedAt: Date;
  
  
  
  
  
  
  
  
  
  pilatDictionaryDicParStatus:PilatParams;
  

  //</es-section>
}